Create README.md
ADW Development
ADW Development is a Scratch Studio, Founded in 2024, That makes anything possible from 3D to 2D, Everything is possible in ADW Development, ADW Development was created by @SwordFish16 (Current CEO) and @DA_FWOGGY_POATO for the purpose to expand the boundaries of Scratch programming to incomprehensible lengths. Ever since it was founded, ADW Development has been expanding, even outside of Scratch. It is continuing to make mind-blowing open-source programs for all.
This is WIP ⛏️⛏️⚒️🪧🪮